Do: Rename directory _Impl_ for _Pages_.
