+++ src/App.tsx
@@ -1,19 +1,19 @@
 import type { RouteObject } from "react-router-dom";
 import { Outlet, Link, useRoutes, useParams } from "react-router-dom";
 
-import Header from './Impl/Header.tsx'
-import Footer from './Impl/Footer.tsx'
-import Menu from './Impl/Menu.tsx'
+import Header from './Pages/Header.tsx'
+import Footer from './Pages/Footer.tsx'
+import Menu from './Pages/Menu.tsx'
 
-import Character_Hulk from './Impl/Characters/Character_Hulk.tsx'
-import Character_IronMan from './Impl/Characters/Character_IronMan.tsx'
-import Character_SpiderMan from './Impl/Characters/Character_SpiderMan.tsx'
-import Screen_Characters from './Impl/Characters/Characters.tsx'
+import Character_Hulk from './Pages/Characters/Character_Hulk.tsx'
+import Character_IronMan from './Pages/Characters/Character_IronMan.tsx'
+import Character_SpiderMan from './Pages/Characters/Character_SpiderMan.tsx'
+import Screen_Characters from './Pages/Characters/Characters.tsx'
 
-import Comics_EssentialIronManVol1 from './Impl/Comics/Comics_EssentialIronManVol1.tsx'
-import Comics_HulkAngry from './Impl/Comics/Comics_HulkAngry.tsx'
-import Comics_ModernSpiderMan from './Impl/Comics/Comics_ModernSpiderMan.tsx'
-import Screen_Comics from './Impl/Comics/Comics.tsx'
+import Comics_EssentialIronManVol1 from './Pages/Comics/Comics_EssentialIronManVol1.tsx'
+import Comics_HulkAngry from './Pages/Comics/Comics_HulkAngry.tsx'
+import Comics_ModernSpiderMan from './Pages/Comics/Comics_ModernSpiderMan.tsx'
+import Screen_Comics from './Pages/Comics/Comics.tsx'
 
 
 function App() {
