+++ src/Pages/Characters/Character_Hulk.tsx
