+++ src/Pages/Characters/Character_SpiderMan.tsx
