+++ src/Pages/Comics/Comics_HulkAngry.tsx
