+++ src/data/objects/Characters.tsx
@@ -2,19 +2,16 @@ export const characters = [
     {
         id: 2,
         name: 'Iron-Man',
-        url: '/characters/2',
         img: 'src/data/image/iron_man.png'
     },
     {
         id: 3,
         name: 'Spider-Man',
-        url: 'characters/2',
         img: 'src/data/image/spider_man.png'
     },
     {
         id: 1,
         name: 'Halk',
-        url: '/characters/1',
         img: 'src/data/image/hulk.png'
     }
 ];
